MdeModulePkg: Fix comparison with wider widths.
https://codeql.github.com/codeql-query-help/cpp/cpp-comparison-with-wider-type

If the narrow type (smaller range) is compared against a wide type
(larger range), the narrow value may overflow before reaching the wide
value. This can cause unexpected behavior, such as:

Infinite loops (loop condition never becomes false).
Incorrect logic (comparison results are misleading).

+++ MdeModulePkg/Bus/Pci/SdMmcPciHcDxe/SdMmcPciHci.c
@@ -1476,7 +1476,7 @@ BuildAdmaDescTable (
   EFI_PHYSICAL_ADDRESS  Data;
   UINT64                DataLen;
   UINT64                Entries;
-  UINT32                Index;
+  UINT64                Index;
   UINT64                Remaining;
   UINT64                Address;
   UINTN                 TableSize;
